recvfile.dpp works
it uses recvfile2. recvfile2 takes a filename as const(char)* which we
can pass easily. recvfile on the other hand takes a c++ fstream which is
a lot harder to construct.
it might be possible to reimplement recvfile with D File, but
I'm not sure if that is worth it.